VideoPrivacyBackend/requirements.txt
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-cloud-videointelligence==2.13.3
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mutagen==1.47.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.20.0
pytube==15.0.0
requests==2.32.3
rsa==4.9
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
youtube-dl==2021.12.17
yt-dlp==2024.8.1
